@@ -30,14 +30,16 @@ class StringBufferWrapper with ChangeNotifier {
 class ShortFormatter extends LogRecordFormatter {
   @override
   StringBuffer formatToStringBuffer(LogRecord rec, StringBuffer sb) {
-    sb.write('${rec.time.hour}:${rec.time.minute}:${rec.time.second} ${rec.level.name} '
+    sb.write(
+        '${rec.time.hour}:${rec.time.minute}:${rec.time.second} ${rec.level.name} '
         '${rec.message}');
 
     if (rec.error != null) {
       sb.write(rec.error);
     }
     // ignore: avoid_as
-    final stackTrace = rec.stackTrace ?? (rec.error is Error ? (rec.error as Error).stackTrace : null);
+    final stackTrace = rec.stackTrace ??
+        (rec.error is Error ? (rec.error as Error).stackTrace : null);
     if (stackTrace != null) {
       sb.write(stackTrace);
     }
@@ -63,13 +65,14 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   final String baseName = 'default';
-  SecureBiometricStorageFile _authStorage;
-  SecureBiometricStorageFile _storage;
-  SecureBiometricStorageFile _customPrompt;
-  SecureBiometricStorageFile _noConfirmation;
-  List<BiometricType> _availableBiometrics;
+  SecureBiometricStorageFile? _authStorage;
+  SecureBiometricStorageFile? _storage;
+  SecureBiometricStorageFile? _customPrompt;
+  SecureBiometricStorageFile? _noConfirmation;
+  List<BiometricType>? _availableBiometrics;
 
-  final TextEditingController _writeController = TextEditingController(text: 'Lorem Ipsum');
+  final TextEditingController _writeController =
+      TextEditingController(text: 'Lorem Ipsum');
 
   @override
   void initState() {
@@ -123,7 +126,8 @@ class _MyAppState extends State<MyApp> {
             const Text('Methods:'),
             ElevatedButton(
               onPressed: () async {
-                final availableBiometrics = await SecureBiometricStorage().getAvailableBiometrics();
+                final availableBiometrics =
+                    await SecureBiometricStorage().getAvailableBiometrics();
                 setState(() {
                   _availableBiometrics = availableBiometrics;
                 });
@@ -133,10 +137,15 @@ class _MyAppState extends State<MyApp> {
             ...?(_availableBiometrics == null
                 ? null
                 : [
-                    const Text('Available Hardware', style: TextStyle(fontWeight: FontWeight.bold)),
+                    const Text('Available Hardware',
+                        style: TextStyle(fontWeight: FontWeight.bold)),
                     Row(
                       mainAxisAlignment: MainAxisAlignment.spaceAround,
-                      children: <Widget>[..._availableBiometrics.map((e) => _mapBiometricType(e)).toList()],
+                      children: <Widget>[
+                        ..._availableBiometrics!
+                            .map((e) => _mapBiometricType(e))
+                            .toList()
+                      ],
                     ),
                   ]),
             const Divider(),
@@ -148,31 +157,40 @@ class _MyAppState extends State<MyApp> {
                 bool supportsAuthenticated = false;
                 if (authenticate == CanAuthenticateResponse.success) {
                   supportsAuthenticated = true;
-                } else if (authenticate != CanAuthenticateResponse.unsupported) {
+                } else if (authenticate !=
+                    CanAuthenticateResponse.unsupported) {
                   supportsAuthenticated = false;
                 } else {
-                  _logger.severe('Unable to use authenticate. Unable to get storage.');
+                  _logger.severe(
+                      'Unable to use authenticate. Unable to get storage.');
                   return;
                 }
                 if (supportsAuthenticated) {
-                  _authStorage = await SecureBiometricStorage().getStorage('${baseName}_authenticated',
-                      options: StorageFileInitOptions(authenticationRequired: true));
+                  _authStorage = await SecureBiometricStorage().getStorage(
+                      '${baseName}_authenticated',
+                      options:
+                          StorageFileInitOptions(authenticationRequired: true));
                 }
-                _storage = await SecureBiometricStorage().getStorage('${baseName}_unauthenticated',
-                    options: StorageFileInitOptions(
-                      authenticationRequired: false,
-                    ));
+                _storage = await SecureBiometricStorage()
+                    .getStorage('${baseName}_unauthenticated',
+                        options: StorageFileInitOptions(
+                          authenticationRequired: false,
+                        ));
                 if (supportsAuthenticated) {
-                  _customPrompt = await SecureBiometricStorage().getStorage('${baseName}_customPrompt',
-                      options: StorageFileInitOptions(authenticationRequired: true),
+                  _customPrompt = await SecureBiometricStorage().getStorage(
+                      '${baseName}_customPrompt',
+                      options:
+                          StorageFileInitOptions(authenticationRequired: true),
                       androidPromptInfo: const AndroidPromptInfo(
                         title: 'Custom title',
                         subtitle: 'Custom subtitle',
                         description: 'Custom description',
                         negativeButton: 'Nope!',
                       ));
-                  _noConfirmation = await SecureBiometricStorage().getStorage('${baseName}_customPrompt',
-                      options: StorageFileInitOptions(authenticationRequired: true),
+                  _noConfirmation = await SecureBiometricStorage().getStorage(
+                      '${baseName}_customPrompt',
+                      options:
+                          StorageFileInitOptions(authenticationRequired: true),
                       androidPromptInfo: const AndroidPromptInfo(
                         confirmationRequired: false,
                       ));
@@ -184,29 +202,41 @@ class _MyAppState extends State<MyApp> {
             ...(_authStorage == null
                 ? []
                 : [
-                    const Text('Biometric Authentication', style: TextStyle(fontWeight: FontWeight.bold)),
-                    StorageActions(storageFile: _authStorage, writeController: _writeController),
+                    const Text('Biometric Authentication',
+                        style: TextStyle(fontWeight: FontWeight.bold)),
+                    StorageActions(
+                        storageFile: _authStorage!,
+                        writeController: _writeController),
                     const Divider(),
                   ]),
             ...?(_storage == null
                 ? null
                 : [
-                    const Text('Unauthenticated', style: TextStyle(fontWeight: FontWeight.bold)),
-                    StorageActions(storageFile: _storage, writeController: _writeController),
+                    const Text('Unauthenticated',
+                        style: TextStyle(fontWeight: FontWeight.bold)),
+                    StorageActions(
+                        storageFile: _storage!,
+                        writeController: _writeController),
                     const Divider(),
                   ]),
             ...?(_customPrompt == null
                 ? null
                 : [
-                    const Text('Custom Authentication Prompt (Android)', style: TextStyle(fontWeight: FontWeight.bold)),
-                    StorageActions(storageFile: _customPrompt, writeController: _writeController),
+                    const Text('Custom Authentication Prompt (Android)',
+                        style: TextStyle(fontWeight: FontWeight.bold)),
+                    StorageActions(
+                        storageFile: _customPrompt!,
+                        writeController: _writeController),
                     const Divider(),
                   ]),
             ...?(_noConfirmation == null
                 ? null
                 : [
-                    const Text('No Confirmation Prompt (Android)', style: TextStyle(fontWeight: FontWeight.bold)),
-                    StorageActions(storageFile: _noConfirmation, writeController: _writeController),
+                    const Text('No Confirmation Prompt (Android)',
+                        style: TextStyle(fontWeight: FontWeight.bold)),
+                    StorageActions(
+                        storageFile: _noConfirmation!,
+                        writeController: _writeController),
                   ]),
             const Divider(),
             ElevatedButton(
@@ -246,7 +276,9 @@ class _MyAppState extends State<MyApp> {
 }
 
 class StorageActions extends StatelessWidget {
-  const StorageActions({Key key, @required this.storageFile, @required this.writeController}) : super(key: key);
+  const StorageActions(
+      {Key? key, required this.storageFile, required this.writeController})
+      : super(key: key);
 
   final SecureBiometricStorageFile storageFile;
   final TextEditingController writeController;
@@ -268,7 +300,8 @@ class StorageActions extends StatelessWidget {
           child: const Text('write'),
           onPressed: () async {
             _logger.fine('Going to write...');
-            await storageFile.write(' [${DateTime.now()}] ${writeController.text}');
+            await storageFile
+                .write(' [${DateTime.now()}] ${writeController.text}');
             _logger.info('Written content.');
           },
         ),
